+++ src/main/java/com/flycms/module/share/model/ShareCategory.java
@@ -21,9 +21,9 @@ import java.io.Serializable;
 public class ShareCategory implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    private Integer id;
+    private Long id;
 
-    private Integer fatherId;
+    private Long fatherId;
 
     //分类所有归属id
     private String categoryId;
